plat/arm: scmi: Move to drivers/ folder
