+++ drivers/arm/css/scmi/scmi_private.h
