+++ include/drivers/arm/css/scmi.h
