+++ plat/arm/board/juno/juno_pm.c
@@ -4,9 +4,8 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
+#include <drivers/arm/css/scmi.h>
 #include <plat/arm/common/plat_arm.h>
-
-#include <scmi.h>
 
 const plat_psci_ops_t *plat_arm_psci_override_pm_ops(plat_psci_ops_t *ops)
 {
